add Layout Management Provider to sample apps
