+++ vuu-ui/packages/vuu-layout/src/layout-persistence/index.ts
@@ -1,3 +1,4 @@
-export * from './LayoutPersistenceManager';
-export * from './LocalLayoutPersistenceManager';
-export * from './data';
+export * from "./data";
+export * from "./LayoutPersistenceManager";
+export * from "./LocalLayoutPersistenceManager";
+export * from "./useLayoutContextMenuItems";
