+++ vuu-ui/packages/vuu-shell/src/shell.tsx
@@ -13,10 +13,7 @@ import {
   LayoutProvider,
   LayoutProviderProps,
 } from "@finos/vuu-layout";
-import {
-  LayoutChangeHandler,
-  LayoutJSON,
-} from "@finos/vuu-layout/src/layout-reducer";
+import { LayoutChangeHandler } from "@finos/vuu-layout/src/layout-reducer";
 import { AppHeader } from "./app-header";
 import { ThemeMode, ThemeProvider, useThemeAttributes } from "./theme-provider";
 import { logger } from "@finos/vuu-utils";
@@ -64,7 +61,8 @@ export const Shell = ({
 }: ShellProps) => {
   const rootRef = useRef<HTMLDivElement>(null);
   const layoutId = useRef("latest");
-  const { applicationLayout, saveApplicationLayout, loadLayoutById } = useLayoutManager();
+  const { applicationLayout, saveApplicationLayout, loadLayoutById } =
+    useLayoutManager();
 
   const handleLayoutChange = useCallback<LayoutChangeHandler>(
     (layout, layoutChangeReason) => {
@@ -76,7 +74,7 @@ export const Shell = ({
         error?.("Failed to save layout");
       }
     },
-    [applicationLayout]
+    [saveApplicationLayout]
   );
 
   const handleSwitchTheme = useCallback((mode: ThemeMode) => {
@@ -123,6 +121,7 @@ export const Shell = ({
   return (
     <ThemeProvider>
       <LayoutProvider
+        {...LayoutProps}
         layout={applicationLayout}
         onLayoutChange={handleLayoutChange}
       >
