+++ vuu-ui/sample-apps/app-vuu-example/index.tsx
@@ -1,7 +1,11 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { App } from "./src/App";
-import { getAuthDetailsFromCookies, redirectToLogin } from "@finos/vuu-shell";
+import {
+  getAuthDetailsFromCookies,
+  LayoutManagementProvider,
+  redirectToLogin,
+} from "@finos/vuu-shell";
 
 import "@salt-ds/theme/index.css";
 import "@finos/vuu-icons/index.css";
@@ -12,7 +16,9 @@ if (!username || !token) {
   redirectToLogin();
 } else {
   ReactDOM.render(
-    <App user={{ username, token }} />,
+    <LayoutManagementProvider>
+      <App user={{ username, token }} />
+    </LayoutManagementProvider>,
     document.getElementById("root")
   );
 }
