+++ vuu-ui/sample-apps/app-vuu-example/src/App.css
@@ -64,3 +64,10 @@ body {
 .vuuToolbarProxy-vertical {
   flex-direction: column;
 }
+
+.vuuShell-mainTabs {
+  height: 100%;
+  position: relative;
+  width: 100%;
+}
+
