+++ vuu-ui/tsconfig.json
@@ -15,7 +15,7 @@
     "module": "esnext",
     "moduleResolution": "node",
     "resolveJsonModule": true,
-    "noEmit": true,
+    // "noEmit": true,
     "jsx": "react-jsx",
     "sourceMap": true
   },
